LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/the-location-interface/per-global.window-expected.txt


PASS Discarding the browsing context must not change window.location
FAIL Navigating from the initial about:blank must not replace window.location assert_equals: expected object "about:blank" but got object "http://localhost:8800/common/blank.html"
PASS document.open() must replace window.location
